Factor out common loaders for select boxes into mixin
